ANDROID: dm-crypt: Remove WQ_NON_REENTRANT flag.
Fix the build of drivers/md/dm-crypt.c.

In 82c6fea "ANDROID: dm-crypt: run in a WQ_HIGHPRI workqueue"
a change was cherry-picked from an earlier kernel which had
the WQ_NON_REENTRANT flag. As per Documentation/workqueue.txt,
this flag has been removed from the kernel and all WQs are
non-reentrant. It is not appropriate for 3.18.

Change-Id: I4837cef258afb7d98c3e4cb140d0506b2902b5cf
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>