qcom: rpm-smd: Increase wait timeout during sleep set
RPM driver waits for 5 us when the the tx buffer is full and eventually
times out after 250us. This isn't sufficient time for RPM to consume the
data. This could result in a unsuccessful attempts to send sleep set
buffers.

Increase the timeout to 100us to give RPM sufficient time to read through
the data.

CRs-fixed: 980952
Change-Id: I4536c2503bbcba3aad3401ca5bace8d1884e3f50
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Signed-off-by: Alex Naidis <alex.naidis@linux.com>